Rename depth_average to compute_average to reflect multiple axis support and update refrences.
